+++ keylogger_module/__init__.py
@@ -1,0 +1,2 @@
+from .keylogger_basics import *
+import pynput.keyboard as keyboard
